refactor(rtc_data_enclave::ocalls): extract mod save_sealed_blob_impl
