+++ rtc_data_enclave/src/ocalls/mod.rs
@@ -1,34 +1,7 @@
-use std::boxed::Box;
+//! OCALL definitions
 
-use sgx_types::*;
-use uuid::Uuid;
+mod save_sealed_blob_impl;
 
-extern "C" {
-    // rtc_save_sealed_blob_u(sgx_status_t* retval, const uint8_t* blob_ptr, size_t blob_len);
-    fn rtc_save_sealed_blob_u(
-        retval: *mut sgx_status_t,
-        blob_ptr: *const u8,
-        blob_len: usize,
-        uuid: &[u8; 16],
-    ) -> sgx_status_t;
-}
+// Re-export the OCALL entry points we're interested in:
 
-pub fn save_sealed_blob_u(blob: Box<[u8]>, uuid: Uuid) -> sgx_status_t {
-    let mut retval = sgx_status_t::default();
-    let sgx_res = unsafe {
-        rtc_save_sealed_blob_u(
-            &mut retval as *mut _,
-            blob.as_ptr(),
-            blob.len(),
-            uuid.as_bytes(),
-        )
-    };
-
-    if retval != sgx_status_t::SGX_SUCCESS {
-        retval
-    } else if sgx_res != sgx_status_t::SGX_SUCCESS {
-        sgx_res
-    } else {
-        sgx_status_t::SGX_SUCCESS
-    }
-}
+pub(crate) use save_sealed_blob_impl::save_sealed_blob_u;
